// src/app/shared/components/form/form-types.ts
export interface McvFieldStyles {
    borderStyle?: string;
    outline?: string;
    textColor?: string;
    backgroundColor?: string;
    activeBorderStyle?: string;
    activeOutline?: string;
    activeTextColor?: string;
    activeBackgroundColor?: string;
    sizeVariant?: 'sm' | 'md' | 'lg';
    // Component specific
    selectedColor?: string;
    labelColor?: string;
    thumbColor?: string;
    trackColor?: string;
}